fastapi/requirements.txt
fastapi
google
google.generativeai
llama-cpp-python
numpy
opencv-python
pandas
pillow
pydantic
pymysql
scikit-learn
sentence-transformers
setproctitle
uvicorn